+++ tests/insertion.rs
@@ -404,6 +404,101 @@ fn after_removal() {
         .single(client_app.world());
 }
 
+#[test]
+fn before_started_replication() {
+    let mut server_app = App::new();
+    let mut client_app = App::new();
+    for app in [&mut server_app, &mut client_app] {
+        app.add_plugins((
+            MinimalPlugins,
+            RepliconPlugins.set(ServerPlugin {
+                tick_policy: TickPolicy::EveryFrame,
+                replicate_after_connect: false,
+                ..Default::default()
+            }),
+        ))
+        .replicate::<TableComponent>();
+    }
+
+    server_app.connect_client(&mut client_app);
+
+    server_app.world_mut().spawn((Replicated, TableComponent));
+
+    server_app.update();
+    server_app.exchange_with_client(&mut client_app);
+    client_app.update();
+    server_app.exchange_with_client(&mut client_app);
+
+    let replicated_components = client_app
+        .world_mut()
+        .query_filtered::<(), With<TableComponent>>()
+        .iter(client_app.world())
+        .count();
+
+    assert_eq!(
+        replicated_components, 0,
+        "no entities should have been sent to the client"
+    );
+
+    let client = client_app.world().resource::<RepliconClient>();
+    let client_id = client.id().unwrap();
+    server_app
+        .world_mut()
+        .send_event(StartReplication(client_id));
+
+    server_app.update();
+    server_app.exchange_with_client(&mut client_app);
+    client_app.update();
+    server_app.exchange_with_client(&mut client_app);
+
+    client_app
+        .world_mut()
+        .query_filtered::<(), With<TableComponent>>()
+        .single(client_app.world());
+}
+
+#[test]
+fn after_started_replication() {
+    let mut server_app = App::new();
+    let mut client_app = App::new();
+    for app in [&mut server_app, &mut client_app] {
+        app.add_plugins((
+            MinimalPlugins,
+            RepliconPlugins.set(ServerPlugin {
+                tick_policy: TickPolicy::EveryFrame,
+                replicate_after_connect: false,
+                ..Default::default()
+            }),
+        ))
+        .replicate::<TableComponent>();
+    }
+
+    server_app.connect_client(&mut client_app);
+
+    let client = client_app.world().resource::<RepliconClient>();
+    let client_id = client.id().unwrap();
+    server_app
+        .world_mut()
+        .send_event(StartReplication(client_id));
+
+    server_app.update();
+    server_app.exchange_with_client(&mut client_app);
+    client_app.update();
+    server_app.exchange_with_client(&mut client_app);
+
+    server_app.world_mut().spawn((Replicated, TableComponent));
+
+    server_app.update();
+    server_app.exchange_with_client(&mut client_app);
+    client_app.update();
+    server_app.exchange_with_client(&mut client_app);
+
+    client_app
+        .world_mut()
+        .query_filtered::<(), With<TableComponent>>()
+        .single(client_app.world());
+}
+
 #[derive(Component, Deserialize, Serialize)]
 struct MappedComponent(Entity);